Lorentz.Contracts.ManagedLedger.Doc

Description

Documentation for ManagedLedger contract.

Documentation

contractDoc :: Markdown

transferDoc :: Markdown

approveDoc :: Markdown

approveCASDoc :: Markdown

getAllowanceDoc :: Markdown

getBalanceDoc :: Markdown

getTotalSupplyDoc :: Markdown

setPauseDoc :: Markdown

setAdministratorDoc :: Markdown

getAdministratorDoc :: Markdown

mintDoc :: Markdown

burnDoc :: Markdown

data DTokenNotPausedOnly

Mention that given action can be executed only when transfer-like operations are not paused.

Constructors

DTokenNotPausedOnly

data DRequireRole

Mention that entrypoint requires sender to have special privileges.

You provide textual description of these privileges, e.g. "administrator" or "trusted".

Constructors

DRequireRole Text