Lorentz.Contracts.Spec.ApprovableLedgerInterface

Description

FA1.2 interface specification.

Documentation

data Parameter

Provides an FA1.2 interface that is supposed to be used by applications that can work with any FA1.2-compatible contract. This type should be transformed to a concrete contract parameter and is not supposed be used in contract code, thus it doesn't derive an IsoValue.

Constructors

Transfer TransferParams

Approve ApproveParams

GetAllowance (View GetAllowanceParams Natural)

GetBalance (View GetBalanceParams Natural)

GetTotalSupply (View () Natural)

type ParameterC param = ParameterContainsEntryPoints param ["Transfer" :> TransferParams, "Approve" :> ApproveParams, "GetAllowance" :> View GetAllowanceParams Natural, "GetBalance" :> View GetBalanceParams Natural, "GetTotalSupply" :> View () Natural]

type TransferParams = ("from" :! Address, "to" :! Address, "value" :! Natural)

type ApproveParams = ("spender" :! Address, "value" :! Natural)

type GetAllowanceParams = ("owner" :! Address, "spender" :! Address)

type GetBalanceParams = "owner" :! Address

Orphan instances

CustomErrorHasDoc "notEnoughAllowance"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

CustomErrorHasDoc "unsafeAllowanceChange"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

Buildable (CustomError "notEnoughAllowance")

Instance details

Methods

build :: CustomError "notEnoughAllowance" -> Builder

Buildable (CustomError "unsafeAllowanceChange")

Instance details

Methods

build :: CustomError "unsafeAllowanceChange" -> Builder